[EZ] Update TT-Transformer Script To Not Double Buffer Weights
By default, the HF transfromer library will download and load weights in FP32, casting them later to the final dtype.
The issue with this is the implementation is known to be bad; many models distribute weights in natively lower bit types. The underlying implementation will for example keep a copy of every weight in fp32 in memory and bf16 which is bad and causes OOM.

Use torch_dtype="auto" to fix this.